feat: create the reusable banner using slots
Naming slot is applied in most of the parts since slot is more reusable,
customizable than prop. And all of the elements are optional including
background image.

Also, banners are listed on the same view now instead of using tabs. No
need to use tabs for only displaying a few banners each tab.